bot/requirements.txt
pyrogram
tgcrypto
sqlalchemy
mysqlclient
python-dotenv
feedparser
celery
redis
ccxt
matplotlib
pandas
TA-Lib
apscheduler
